requirements.txt
pandas
streamlit
pyyaml
st-gsheets-connection
altair